use dynamic import for katex and cleanup dependencies
