+++ package.json
@@ -4,20 +4,20 @@
   "description": "Hugo theme made for documentation",
   "main": "index.js",
   "scripts": {
-    "build": "run-s prep svg build:webpack-prod",
-    "build:webpack-prod": "webpack --mode=production",
+    "build": "run-s prep:clean prep:make svg build:webpack",
+    "build:webpack": "webpack --mode=production",
     "build:webpack-devel": "webpack --mode=development",
-    "start": "run-s prep svg build:webpack-devel ; run-p start:webpack start:hugo",
+    "start": "run-s prep:clean prep:make svg build:webpack-devel ; run-p start:webpack start:hugo",
     "start:webpack": "webpack --mode=development --watch",
     "start:build": "webpack --mode=development",
     "start:hugo": "hugo server -D -F -s exampleSite/",
     "svg": "run-s svg:*",
     "svg:sprite": "svg-sprite -C svgsprite.config.json 'src/icons/*.svg'",
     "svg:font": "svgtofont --sources build/icons --output build/fonts",
-    "prep": "run-s prep:*",
-    "prep:clean": "del build/ static/",
-    "prep:make": "make-dir build/icons build/fonts",
-    "svg-sprite-list": "run-s prep:make svg ; make-dir exampleSite/data/sprites ; cpy build/fonts/GeekdocIcons.json exampleSite/data/sprites/ --rename geekdoc.json",
+    "prep:clean": "shx rm -rf build/ static/",
+    "prep:clean-all": "shx rm -rf build/ static/ data/ assets/sprites/ exampleSite/data/sprites",
+    "prep:make": "mkdir -p build/icons build/fonts",
+    "svg-sprite-list": "run-s prep:make svg ; mkdir -p exampleSite/data/sprites/ ; cp build/fonts/GeekdocIcons.json exampleSite/data/sprites/geekdoc.json",
     "lint": "eslint src/js/ --color"
   },
   "repository": {
@@ -38,20 +38,18 @@
     "@babel/eslint-parser": "7.16.5",
     "autoprefixer": "10.4.1",
     "copy-webpack-plugin": "10.2.0",
-    "cpy-cli": "3.1.1",
     "css-loader": "6.5.1",
-    "del-cli": "4.0.1",
     "eslint": "8.5.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "4.0.0",
     "favicons-webpack-plugin": "5.0.2",
-    "make-dir-cli": "3.0.0",
     "npm-run-all": "4.1.5",
     "postcss-loader": "6.2.1",
     "prettier": "2.5.1",
     "prettier-plugin-go-template": "0.0.11",
     "sass": "1.45.1",
     "sass-loader": "12.4.0",
+    "shx": "0.3.3",
     "svg-sprite": "1.5.3",
     "svgtofont": "3.13.1",
     "webpack": "5.65.0",
